Fix tests via moving dag test related test to local_command
